requirements.txt
fastapi
uvicorn
pydantic
pandas
openpyxl
google-generativeai